requirements.txt
algoliasearch-django
django
djangorestframework
pyyaml
requests
django-cors-headers
